+++ src/Raven.CodeAnalysis/Binder/FunctionExpressionBinder.cs
@@ -10,7 +10,6 @@ class FunctionExpressionBinder : BlockBinder
     private readonly Dictionary<string, IParameterSymbol> _parameters = new();
     private readonly Dictionary<string, IMethodSymbol> _functions = new();
     private readonly List<ILocalSymbol> _declaredLocals = [];
-    private ImplicitReceiverSymbol? _implicitReceiver;
 
     public FunctionExpressionBinder(ISymbol containingSymbol, Binder parent) : base(containingSymbol, parent) { }
 
@@ -23,20 +22,6 @@ class FunctionExpressionBinder : BlockBinder
     }
 
     public IEnumerable<IParameterSymbol> GetParameters() => _parameters.Values;
-
-    public void DeclareImplicitReceiverParameter(IParameterSymbol parameter, ITypeSymbol? lookupType = null)
-    {
-        _implicitReceiver = new ImplicitReceiverSymbol(parameter, lookupType ?? parameter.Type);
-    }
-
-    public void DeclareImplicitReceiverLocal(ILocalSymbol local, ITypeSymbol lookupType)
-    {
-        _implicitReceiver = new ImplicitReceiverSymbol(local, lookupType);
-    }
-
-    public ImplicitReceiverSymbol? GetImplicitReceiver() => _implicitReceiver;
-
-    public IParameterSymbol? GetImplicitReceiverParameter() => _implicitReceiver?.Symbol as IParameterSymbol;
 
     public void DeclareFunction(string name, IMethodSymbol method)
     {
@@ -118,12 +103,6 @@ class FunctionExpressionBinder : BlockBinder
     {
         if (_parameters.Values.Contains(symbol))
             return true;
-
-        if (_implicitReceiver is { } implicitReceiver &&
-            SymbolEqualityComparer.Default.Equals(implicitReceiver.Symbol, symbol))
-        {
-            return true;
-        }
 
         if (symbol is IParameterSymbol parameter &&
             _parameters.Values.Any(declaredParameter => HaveSameDeclaration(declaredParameter, parameter)))